+++ docs/source/changelog.rst
@@ -11,16 +11,22 @@ Versions
 latest
 ~~~~~~
 
+Notable enhancements and changes are:
+
+    * :issue:`81` - :func:`pywincffi.user32.synchronization.WSAEventSelect` and
+      :func:`pywincffi.user32.synchronization.WSAEnumNetworkEvents`
+
 0.3.0
 ~~~~~
 
 Notable enhancements and changes are:
 
-    * Added the :func:`pywincffi.kernel32.CreateToolhelp32Snapshot` function in :issue:`101`.
+    * Added the :func:`pywincffi.kernel32.CreateToolhelp32Snapshot` function
+      in :issue:`101`.
     * Fixed a bug where :func:`pywincffi.checks.input_check` might raise
       ``ffi.error`` in :issue:`73`
     * Several enhancements bringing :issue:`69` closer to closure.
-    * Addition several functions or :issue:`69`:
+    * Addition several functions for :issue:`69`:
         * :issue:`70` - :func:`pywincffi.kernel32.events.CreateEvent` and
           :func:`pywincffi.kernel32.events.OpenEvent`
         * :issue:`75` - :func:`pywincffi.kernel32.events.ResetEvent`